Use FileImageSourcePathConverter on NavigationViewItem instead of a custom control (so I deleted ShellNavigationViewItemRenderer which is no longer needed). Ensure `FileImageSourcePathConverter` won't throw if it didn't get a FileImageSource. Move the flyout data templates into a resource so they can be overridden and compiled.
