+++ Xamarin.Forms.Platform.UAP/Resources.xaml
@@ -14,6 +14,8 @@
 		<ResourceDictionary Source="MasterDetailControlStyle.xaml" />
 		<ResourceDictionary Source="TabbedPageStyle.xaml" />
 		<ResourceDictionary Source="SliderStyle.xaml" />
+		<ResourceDictionary Source="Items/ItemsViewStyles.xaml" />
+        <ResourceDictionary Source="Shell/ShellFlyoutTemplateSelectorStyles.xaml" />
 		<ResourceDictionary Source="CollectionView/ItemsViewStyles.xaml" />
 	</ResourceDictionary.MergedDictionaries>
